backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
scrapy==2.11.0
beautifulsoup4==4.12.2
openai==1.3.7
pinecone==7.3.0
python-dotenv==1.0.0
pydantic==2.4.2
requests==2.31.0
lxml==4.9.3
python-multipart==0.0.6
httpx==0.25.2 